Moved Login / logout functions to useLogin hook
